refactor: Badge index.jsx 생성
