+++ fe/src/Components/Badge/index.jsx
@@ -1,0 +1,2 @@
+export { CategoryBadge } from './CategoryBadge'
+export { DiscountBadge } from './DiscountBadge'
